% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{adjustCatVarCalib}
\alias{adjustCatVarCalib}
\title{Adjust categorical values to desired proportions in cat.adjustments (if any).}
\usage{
adjustCatVarCalib(x, varname, propens = NULL, desiredProps = NULL,
  simenv = simenv, iteration = iteration)
}
\arguments{
\item{x}{categorical values to adjust}

\item{varname}{varname, used a lookup into cat.adjustments and propensities}
}
\description{
Allows subgroup adjustment if a subgroup expression attribute is attached to the cat.adjustments.
}

