% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{adjustCatVarSimple}
\alias{adjustCatVarSimple}
\title{Adjust categorical values to desired proportions in cat.adjustments (if any).}
\usage{
adjustCatVarSimple(x, varname, simenv = simenv, iteration = iteration)
}
\arguments{
\item{x}{categorical values to adjust}

\item{varname}{varname, used a lookup into cat.adjustments and propensities}
}
\description{
Does not allow subgroup adjustments.
}
\examples{
 varname <- "z1msmokeLvl1"
 varname <- "z1singleLvl1"
 adjustCatVar(predSimBinomsSelect(z1single_previousLvl1, models$z1singlePrev0, models$z1singlePrev1), "z1singleLvl1")
}

