% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{adjustContVar}
\alias{adjustContVar}
\title{Adjust continuous values to desired proportions in cat.adjustments (if any).}
\usage{
adjustContVar(
  x,
  varname,
  simenv = simenv,
  propens = NULL,
  desiredProps = NULL,
  iteration
)
}
\arguments{
\item{x}{continuous values to adjust}

\item{varname}{varname, used a lookup into cat.adjustments and propensities}
}
\description{
Allows subgroup adjustment if a global subgroup expression is set.
}
