% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{adjustContVarSimple}
\alias{adjustContVarSimple}
\title{Adjust continuous values to desired proportions in cat.adjustments (if any).}
\usage{
adjustContVarSimple(x, varname, simenv, iteration)
}
\arguments{
\item{x}{continuous values to adjust}

\item{varname}{varname, used a lookup into cat.adjustments and propensities}
}
\description{
Does not allow subgroup adjustment
}
\examples{
 
}

