% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{align.by.name.list.mx}
\alias{align.by.name.list.mx}
\title{Align a list of matrices/vectors by row and column name, so all matrices have the 
same set of named rows and named cols. 
NAs are used to pad missing rows and cols.}
\usage{
align.by.name.list.mx(listmx)
}
\arguments{
\item{listmx}{a list of different sized matrices}
}
\value{
the list of the matrices aligned to the maximal set of rows and cols.
 rows and cols are in numerical sort order.
}
\description{
Align a list of matrices/vectors by row and column name, so all matrices have the 
same set of named rows and named cols. 
NAs are used to pad missing rows and cols.
}
