% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{append.list.mx}
\alias{append.list.mx}
\title{Append two lists of matrices by adding each element of listmx.src into 
the corresponding element of listmx.dest in the specified dimension.}
\usage{
append.list.mx(listmx.dest, listmx.src, into.dim = ZDIM)
}
\arguments{
\item{listmx.dest}{list of vector/matrix/arrays}

\item{listmx.src}{list of vector/matrix/arrays}

\item{into.dim}{the dimension in which to add each element of listmx.src.

 If into.dim = ZDIM (i.e: 3) then listmx.src elements will be added as an additional
 z matrix into each matrix/array element of listmx.dest. Existing elements of listmx.dest
 will be in the first z dim(s). 

 If into.dom = COL (i.e: 2) then listmx.src elements will be added as the 2nd
 col into each vector/matrix element of listmx.dest. Existing elements of listmx.dest
 will be in the first column(s).

 If into.dom = ROW (i.e: 1) then listmx.src elements will be added as the 2nd
 row into each vector/matrix element of listmx.dest. Existing elements of listmx.dest
 will be in the first row(s).}
}
\value{
list of vector/matrix/arrays after appended.
}
\description{
Append two lists of matrices by adding each element of listmx.src into 
the corresponding element of listmx.dest in the specified dimension.
}
