% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{applyAllCatAdjustmentsToSimframe}
\alias{applyAllCatAdjustmentsToSimframe}
\title{Apply categorical adjustments to simframe.}
\usage{
applyAllCatAdjustmentsToSimframe(Simenv, iteration, propensities = NULL,
  print_adj = TRUE, cat.adjustments = Simenv$cat.adjustments)
}
\arguments{
\item{Simenv}{simenv receiving object. Simenv$simframe is modified.}

\item{iteration}{iteration number - corresponds to a row number in the matrix elements of the cat.adjustments list}

\item{propensities}{named list of propensities for the cat.adjustments}

\item{cat.adjustments}{a list of categorical adjustment matrices whose rows each correpond to desired adjustments for an iteration.
   Each matrix has a 'varname' attribute, indicating which variable in the simenv object is to be adjusted.
   Each matrix may also have a 'logisetexpr' attribute - if so, this is evaluated and becomes a logical vector indicating which observations
of the 'varname' variable to adjust (i.e. the "logisetexpr" attribute gives which subset of the data the row of adjustments are intended for).}

\item{printAdj}{if TRUE will print new proportions of modified simframe vars}
}
\value{
NULL. simframe in receiving object is modified directly.
}
\description{
Apply categorical adjustments to simframe.
}

