% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{applyCatAdjustmentToSimframe}
\alias{applyCatAdjustmentToSimframe}
\title{Apply categorical adjustments to simframe.}
\usage{
applyCatAdjustmentToSimframe(Simenv, varnames, desired_props, iteration,
  propensities, print_adj = TRUE)
}
\arguments{
\item{Simenv}{simenv receiving object. Simenv$simframe is modified.}

\item{varnames}{varname(s) of variable(s) to adjust, eg: "catpregsmk2" or c("z1msmokeLvl0","z1msmokeLvl1")}

\item{desired_props}{a vector of desired proportions, eg: c(0.1, 0.1, 0.8).
Can have a "logisetexpr" attribute - if so, this is evaulated, and becomes a logical vector indicating which observations of "varname" to adjust.
       (i.e. the "logisetexpr" attribute gives which subset of the data the desired_props are intended for).}

\item{iteration}{the current iteration}

\item{propensities}{named list of propensities for the cat.adjustments}

\item{printAdj}{if TRUE will print new proportions of modified simframe vars}
}
\value{
NULL. simframe in receiving object is modified directly.
}
\description{
Apply categorical adjustments to simframe.
}

