% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{applyCatAdjustmentToSimframeVarMultipleBinary}
\alias{applyCatAdjustmentToSimframeVarMultipleBinary}
\title{Adjust the proportions of a simframe variable that exists in multiple binary level vectors,
eg: SESBTHLvl1, SESBTHLvl2, SESBTHLvl3.}
\usage{
applyCatAdjustmentToSimframeVarMultipleBinary(
  Simenv,
  binLevelVarnames,
  desiredProps,
  propens,
  printAdj = TRUE,
  logiset = NULL
)
}
\arguments{
\item{Simenv}{simenv receiving object. Simenv$simframe is modified.}

\item{binLevelVarnames}{vector of binary level varnames, eg: c("z1accomLvl0","z1accomLvl1")}

\item{desiredProps}{desired proportions}

\item{propens}{propensities, if ANY}

\item{printAdj}{if TRUE, display adjusted proportions after adjustment}

\item{logiset}{logical vector indicating which rows to include, or NULL (the default) to include all.}
}
\value{
NULL. simframe in receiving object is modified directly.
}
\description{
Adjust the proportions of a simframe variable that exists in multiple binary level vectors,
eg: SESBTHLvl1, SESBTHLvl2, SESBTHLvl3.
}
