% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{applyCatAdjustmentToSimframeVarSingle}
\alias{applyCatAdjustmentToSimframeVarSingle}
\title{Adjust the proportions of a single simframe variable.}
\usage{
applyCatAdjustmentToSimframeVarSingle(Simenv, varname, desired_props, propens,
  print_adj = T, logiset = NULL)
}
\arguments{
\item{Simenv}{simenv receiving object. Simenv$simframe is modified.}

\item{varname}{simframe variable to adjust}

\item{desired_props}{a vector of desired proportions, eg: c(0.1, 0.1, 0.8)}

\item{propens}{propensities for this variable, if any}

\item{logiset}{logical vector indicating which rows to include, or NULL (the default) to include all.}

\item{printAdj}{if TRUE, display adjusted proportions after adjustment}
}
\value{
NULL. simframe in receiving object is modified directly.
}
\description{
Adjust the proportions of a single simframe variable.
}

