% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{as.arrayFromList}
\alias{as.arrayFromList}
\title{convert list of vectors to an array.
Recycling rule will be applied.}
\usage{
as.arrayFromList(mylist)
}
\arguments{
\item{mylist}{list of vectors}
}
\value{
an array
}
\description{
convert list of vectors to an array.
Recycling rule will be applied.
}
