% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{as.matrixFromList}
\alias{as.matrixFromList}
\title{Converts a list of same length vectors to a matrix.
Unlike as.matrix.default, it unlists first and
uses names from the first list element in the result.}
\usage{
as.matrixFromList(xlist, byrow = TRUE)
}
\arguments{
\item{xlist}{a list of same length vectors}

\item{byrow}{if TRUE, each list element becomes a row in the matrix
else, each list element becomes a column in the matrix}
}
\value{
a matrix
}
\description{
Converts a list of same length vectors to a matrix.
Unlike as.matrix.default, it unlists first and
uses names from the first list element in the result.
}
