% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{as_array_list_mx}
\alias{as_array_list_mx}
\title{Convert a list of matrices into an array, with each matrix in the z dimension.}
\usage{
as_array_list_mx(listmx)
}
\arguments{
\item{listmx}{list of matrices. Each must matrix the same dimensions.}
}
\value{
array with dimnames and meta from listmx
}
\description{
Convert a list of matrices into an array, with each matrix in the z dimension.
}
