% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{binary.levels.combine}
\alias{binary.levels.combine}
\title{Combine levels specified in seperate binary level variables into one
variable.}
\usage{
binary.levels.combine(..., levelvalues = NULL)
}
\arguments{
\item{...}{a series of binary (0,1) vectors, specified as multiple arguments
or as a list}

\item{levelvalues}{a vector indicating the level values to use. The first value of this vector is the value used
for the first binary vector argument, the second  for the second binary vector argument etc.
If unspecified, the defaults to (1,2,...n) where n is the number of binary vectors
specified by ...}
}
\value{
a single vector
}
\description{
Combine levels specified in seperate binary level variables into one
variable.
}

