% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{binary.levels.split}
\alias{binary.levels.split}
\title{Create a list of binary levels from factors.
Compares a vector x against a set of factors and returns
a list for each factor which indicates whether x 
is equal to that factor or not.}
\usage{
binary.levels.split(x, f = sort(unique(x)))
}
\arguments{
\item{x}{vector}

\item{f}{factors. Defaults to the unique set of values in x.}
}
\value{
list of binary levels
}
\description{
Create a list of binary levels from factors.
Compares a vector x against a set of factors and returns
a list for each factor which indicates whether x 
is equal to that factor or not.
}
