% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{c_list}
\alias{c_list}
\title{Select inner elements of a list of lists, and return as a new list.}
\usage{
c_list(xlist, fx)
}
\arguments{
\item{xlist}{list}

\item{fx}{function specifing inner elements to select. Should take a single element of xlist
as a parameter.}
}
\value{
equivalent to c(fx(xlist[[1]]), fx(xlist[[2]] ... ) across every element.
}
\description{
Select inner elements of a list of lists, and return as a new list.
}

