% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{check.subgroup.expr}
\alias{check.subgroup.expr}
\title{Check number of valid subgroup while applying cat.adjustment}
\usage{
check.subgroup.expr(Simenv)
}
\arguments{
\item{Simenv}{simenv receiving object. Simenv$simframe is modified.}
}
\value{
numeric number.
}
\description{
Check number of valid subgroup while applying cat.adjustment
}
