% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{cmatch}
\alias{cmatch}
\title{Return the category names for a vector of coded values.
If no category names, returns the list as is.}
\usage{
cmatch(dict, x, varname)
}
\arguments{
\item{x}{coded values}

\item{varname}{name of the category of x, or NULL}

\item{.}{this}
}
\value{
x, is there are no codings for varname, else a vector of category names
 corresponding to the values in x
}
\description{
Return the category names for a vector of coded values.
If no category names, returns the list as is.
}

