% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{cmatchFlattened}
\alias{cmatchFlattened}
\title{Returns the category names for the vector of flattened codes.}
\usage{
cmatchFlattened(dict, x.flat, varname, grpby.tag)
}
\arguments{
\item{dict}{a dictionary object}

\item{x.flat}{a vector of flattened codes. A flattened code is in the form "0 1",
where the first value is a grping code and the second a varname code.
If grpby.tag is NULL or NA, then the flattened code will be in the form "0", 
i.e: no grping codes only varname codes.}

\item{varname}{identifies the varname coding}

\item{grpby.tag}{identifies the grping coding, or NULL if no grouping coding.}
}
\value{

}
\description{
Returns the category names for the vector of flattened codes.
}
