% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{cnamesLookup}
\alias{cnamesLookup}
\title{Returns the names of the codings for the supplied variables.
NB: Requires codings variable to have been set in global environment.}
\usage{
cnamesLookup(dict, vars)
}
\arguments{
\item{dict}{a dictionary object}

\item{vars}{vector of variable names}
}
\value{
list of coding names for the variables supplied
}
\description{
Returns the names of the codings for the supplied variables.
NB: Requires codings variable to have been set in global environment.
}
