% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{collate_all_run_results}
\alias{collate_all_run_results}
\title{Collates (ie: reduces) all run results to averaged values and labels collated results.}
\usage{
collate_all_run_results(Simmodule, all_run_results, cat.adjustments = NULL,
  simframe)
}
\arguments{
\item{Simmodule}{receiving Simmodule object.}

\item{all_run_results}{the list of all run results generated by map_outcomes_to_run_results()}

\item{cat.adjustments}{the list of outcome matrices generated by simulateRun()}

\item{simframe}{the simulation environment's simframe. Useful for accessing adjusted variables.}
}
\value{
any object. Stored in the run_results_collated list.
}
\description{
Collates (ie: reduces) all run results to averaged values and labels collated results.
}

