% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{createCodings}
\alias{createCodings}
\title{Create the codings of the dictionary object. Called by new().}
\usage{
createCodings(codings_dataframe)
}
\arguments{
\item{codings_dataframe}{}
}
\value{
a list of category names for categorical variables.
}
\description{
Create the codings of the dictionary object. Called by new().
}
