% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{createDescriptions}
\alias{createDescriptions}
\title{Create the descriptions of the dictionary object. Called by new().}
\usage{
createDescriptions(descriptions_dataframe)
}
\arguments{
\item{codings_dataframe}{}
}
\description{
Create the descriptions of the dictionary object. Called by new().
}
