% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{createEmptyCatAdjustments}
\alias{createEmptyCatAdjustments}
\title{Create empty categorical variable adjustment matrices.}
\usage{
createEmptyCatAdjustments(simframe, dict, numiterations = NUM_ITERATIONS)
}
\description{
Create empty categorical variable adjustment matrices.
}
