% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{dimnames_prepend_header}
\alias{dimnames_prepend_header}
\title{Push row/col headers (ie: names of the rownames and colnames) into the rownames and colnames.}
\usage{
dimnames_prepend_header(mx)
}
\arguments{
\item{mx}{matrix}
}
\value{
matrix with prepended dimnames
}
\description{
Push row/col headers (ie: names of the rownames and colnames) into the rownames and colnames.
}

