% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{dlookup}
\alias{dlookup}
\title{Lookup description of variable x in the dictionary
first determines the name of variable x, then does the lookup}
\usage{
dlookup(dict, x)
}
\arguments{
\item{dict}{a dictionary object}

\item{x}{}
}
\value{

}
\description{
Lookup description of variable x in the dictionary
first determines the name of variable x, then does the lookup
}
