% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{dlookup_exists}
\alias{dlookup_exists}
\title{A version of dlookup that can be use to see if a given expression is a variable in 
the dictionary or if it is not.  
Useful when users specifiy their own subgroup expression.  If the expression is not 
a name in the dictionary then one can assume that the user has specified their own 
expression to be used for subgrouping purposes.}
\usage{
dlookup_exists(dict, x)
}
\description{
A version of dlookup that can be use to see if a given expression is a variable in 
the dictionary or if it is not.  
Useful when users specifiy their own subgroup expression.  If the expression is not 
a name in the dictionary then one can assume that the user has specified their own 
expression to be used for subgrouping purposes.
}

