% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileUtil.r
\name{file_extension}
\alias{file_extension}
\title{Returns the file extension, i.e: everything after the final dot.}
\usage{
file_extension(x)
}
\arguments{
\item{x}{chacter vector}
}
\value{
file extension, eg: "csv"
}
\description{
Returns the file extension, i.e: everything after the final dot.
}

