% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{flatten_mx_to_row}
\alias{flatten_mx_to_row}
\title{Flatten a vector or matrix (r,c) into a single row matrix (1,c) 
creating new column names from a combination of old rownames 
and colnames.}
\usage{
flatten_mx_to_row(mx, row.names.first = FALSE)
}
\arguments{
\item{mx}{matrix}

\item{row.names.first}{if TRUE, row.names will appear first in the names of the resultant columns, eg: becomes A1,B1,A2,B2
else row.names will appear last, eg: 1A, 1B, 2A, 2B}
}
\value{
a single row matrix (1,c)
}
\description{
A single row matrix, unlike a vector, has the advantage that 
it can have a row name.
 
eg: A1, A2,
    B1, B2
}
