% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{flatten_mxs_to_single_mx}
\alias{flatten_mxs_to_single_mx}
\title{Takes a list of matrices, flattens them into a single row, aligns them by column name
and then returns a single matrix.}
\usage{
flatten_mxs_to_single_mx(listmx)
}
\arguments{
\item{listmx}{list of matrices to flatten. Can have different dimensions but must have similar names
for alignment (@seealso listmx.flat.aligned)}
}
\value{
a single matrix with 
   colnames = the flattened row/col names of listmx[[1]]],
   meta = meta of listmx[[1]]
   rownames = names of listmx
}
\description{
Takes a list of matrices, flattens them into a single row, aligns them by column name
and then returns a single matrix.
}
