% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{generatePreSimulationStats}
\alias{generatePreSimulationStats}
\title{Generate pre simulation stats after adjustment but before simulation begins.}
\usage{
generatePreSimulationStats(Simenv, simframe)
}
\description{
Typically these will be descriptive statistics of input variables that don?t change eg: gender, ethnicity
}
\details{
Sub-classes override this function.
}

