% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simframe.r
\name{getOutcomeVars}
\alias{getOutcomeVars}
\title{Get a set of outcome variables from the simframe.}
\usage{
getOutcomeVars(simframe, outcome_type_select = NULL,
  outcome_module_name = NULL, sorted = FALSE)
}
\arguments{
\item{simframe}{simframe}

\item{outcome_type_select}{a vector of outcome_types, or NULL to select all types.}

\item{outcome_module_name}{a vector of outcome_module, , or NULL to select all sets.}

\item{sorted}{return outcome variables sorted by outcome name?}
}
\value{
a vector of simframe variables (with names the same as the vector value)
}
\description{
Get a set of outcome variables from the simframe.
}
