% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{incByFactor}
\alias{incByFactor}
\title{Increments x by factor}
\usage{
incByFactor(x, factoredx, factorincrements)
}
\arguments{
\item{x}{numeric vector to be incremented}

\item{factoredx}{x factored}

\item{factorincrements}{amounts to increment each factor of factoredx by}
}
\value{
x incremented by the factor increments
}
\description{
Increments x by factor
}

