% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{is.model.var}
\alias{is.model.var}
\title{Checks which models, in a list of models, contain a specified variable}
\usage{
is.model.var(varname, model.list)
}
\arguments{
\item{varname}{the name of the variable for which the user would like to know in which models it appears.
Must contain the "Lvl" suffix}

\item{models}{a list of glm models}
}
\value{
a vector of model names
}
\description{
Checks which models, in a list of models, contain a specified variable
}

