% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label.r
\name{labelColFromVec}
\alias{labelColFromVec}
\title{set the columns names of each x in xlist to
each char vector c in colnameslist}
\usage{
labelColFromVec(xlist, colnameslist)
}
\arguments{
\item{xlist}{list}

\item{colnameslist}{the list of colnames}
}
\value{
a list with named column
}
\description{
set the columns names of each x in xlist to
each char vector c in colnameslist
}
