% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collate.r
\name{labelColumnCodes}
\alias{labelColumnCodes}
\title{Label columns of a 3D array with the codings of the specified varname.}
\usage{
labelColumnCodes(x, dict, varname)
}
\arguments{
\item{x}{vector/array with a column for each category, ordered}

\item{dict}{a Dictionary proto object}

\item{varname}{categorical variable name. The codings for this variable are applied
as column names.}
}
\value{
x with column names that use the codings of varname, and a column label 
 that is the decsription of varname.
}
\description{
Label columns of a 3D array with the codings of the specified varname.
}

