% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{lapply.args.as.list}
\alias{lapply.args.as.list}
\title{Call lapply with a list of arguments to FUN specified as the parameter FUN.args.}
\usage{
lapply.args.as.list(X, FUN, FUN.args)
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object. Other objects (including classed objects) will be coerced by base::as.list.}

\item{FUN}{the function to be applied to each element of xlist}

\item{FUN.args}{a named list of arguments to be suppplied to FUN when executed on each element of xlist}
}
\description{
Call lapply with a list of arguments to FUN specified as the parameter FUN.args.
}

