% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{lapply.subset.append.lol.args}
\alias{lapply.subset.append.lol.args}
\title{Apply lapply.subset.append to each element of lol, using the arguments from 
the parallel list lol.args and the function .FUN.}
\usage{
lapply.subset.append.lol.args(X, lol, lol.args = attr(lol, "args.list"),
  simplify = TRUE, .FUN)
}
\arguments{
\item{X}{vector, list or dataframe on which to execute .FUN}

\item{lol}{a list of lists containing matrix elements}

\item{lol.args}{For each list in lol, there is a set of arguments that are supplied to .FUN.
Each matrix in the list will use the arguments from its corresponding lol.args list.
e.g: .FUN will be called on matrices in lol[[1]] with arguments supplied in
     lol.args[[1]] 
Defaults to a "args.list" attribute on lol.}

\item{simplify}{if TRUE, then results will be added to the Z dimension of the
matrix elements in lol. If FALSE, then results will be added
as a extra list element to each matrix elements of lol}

\item{.FUN}{.FUN to execute over elements of X}
}
\description{
Apply lapply.subset.append to each element of lol, using the arguments from 
the parallel list lol.args and the function .FUN.
}

