% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{lapply.subset.append}
\alias{lapply.subset.append}
\title{Execute a function over a subset of a list, and merges the results into
an existing list.}
\usage{
lapply.subset.append(
  results.list,
  X,
  indices = names(results.list),
  simplify = TRUE,
  .FUN,
  ...
)
}
\arguments{
\item{results.list}{an existing list, eg:

List of 2
$ gptotvis: num [1:5, 1, 1] 5.82 4.24 4.09 3.93 4.11
    ...
$ hadmtot : num [1:5, 1, 1] 0.2344 0.1014 0.1163 0.0847 0.0921
 ...}

\item{X}{vector, list or dataframe on which to execute .FUN}

\item{indices}{elements of X to apply .FUN to. If NULL, returns results.list.
Defaults to names(results.list).}

\item{simplify}{if TRUE, then results will be added to the Z dimension of the
matrices in results.list. If FALSE, then results will be added
as a extra list element to each element of results.list}

\item{.FUN}{.FUN to execute over elements of X}

\item{...}{additional arguments to .FUN}
}
\description{
Execute a function over a subset of a list, and merges the results into
an existing list.
}
