% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{lapply.subset}
\alias{lapply.subset}
\title{Executes .FUN on the specified varnames in X and return the results as a list.}
\usage{
lapply.subset(X, indices, .FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object. Other objects (including classed objects) 
will be coerced by base::as.list.}

\item{indices}{indices, either numeric or names, of elements in X for which to apply .FUN, or NULL to apply to all elements of X}

\item{.FUN}{function}

\item{...}{optional arguments to .FUN.}
}
\value{
results of .FUN applied to X as a list
}
\description{
Executes .FUN on the specified varnames in X and return the results as a list.
}
