% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{lzip}
\alias{lzip}
\title{Takes a list of lists, and combines the corresponding elements of each list
into a result list of lists. The first list of the result contain the first elements
of the source lists, the second list of the result contains the second elements
of the sources lists, and so on.}
\usage{
lzip(lol)
}
\arguments{
\item{lol}{list of lists. Each inner list should have the same number of elements}
}
\value{
the corresponding elements of lol combined
}
\description{
eg: str(lol)
List of 2
 $ A:List of 3
  ..$ A1: chr "a1"
  ..$ A2: chr "a2"
  ..$ A3: chr "a3"
 $ B:List of 3
  ..$ B1: chr "b1"
  ..$ B2: chr "b2"
  ..$ B3: chr "b3"
}
\details{
The result of zip is:

List of 3
 $ A1:List of 2
  ..$ A: chr "a1"
  ..$ B: chr "b1"
 $ A2:List of 2
  ..$ A: chr "a2"
  ..$ B: chr "b2"
 $ A3:List of 2
  ..$ A: chr "a3"
  ..$ B: chr "b3"

zip is the same as the Lisp operation zip:

(zip '(a1 a2 a3) 
		'(b1 b2 b3)
		=> ((a1 b1) (a2 b2) (a3 b3))
}
