% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{map_outcomes_to_run_resultsP}
\alias{map_outcomes_to_run_resultsP}
\title{Map outcomes to run results for a single run. Run results are typically 
descriptive statistics (eg: freqs, means, etc) generated for variables 
in outcomes.}
\usage{
map_outcomes_to_run_resultsP(run, simframe, outcomes, cat.adjustments)
}
\arguments{
\item{run}{the list of outcome matrices generated by simulateRun()}

\item{simframe}{the simulation environment's simframe. Useful for accessing adjusted variables.}

\item{outcomes}{the list of outcome matrices generated by simulateRun()}

\item{cat.adjustments}{the list of outcome matrices generated by simulateRun()}

\item{moduleName}{the list of outcome matrices generated by simulateRun()}
}
\value{
a list of run results. Each element E1 of run results is a list and is the set of results for a given 
 operation. In turn, each element of E1 may be a list, matrix or vector and is the result for
 the operation applied to an outcome variable (ie: an element in outcomes).
}
\description{
Map outcomes to run results for a single run. Run results are typically 
descriptive statistics (eg: freqs, means, etc) generated for variables 
in outcomes.
}

