% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{mem.lx}
\alias{mem.lx}
\title{Return amount of memory, in bytes, used by each
element of a list}
\usage{
mem.lx(lx)
}
\arguments{
\item{lx}{list}
}
\value{
num vector of bytes per element
}
\description{
Return amount of memory, in bytes, used by each
element of a list
}

