% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{merge_list_mx.by.cols}
\alias{merge_list_mx.by.cols}
\title{Takes any number of 2D matrices, each with the same number of cols but with 
any set of named rows. Every col 1 of the matrices are combined into a new 
matrix, as are every col 2,3 etc.}
\usage{
merge_list_mx.by.cols(...)
}
\arguments{
\item{...}{matrices, or a list of matrices}
}
\value{
matrices, or a list of matrices after merged
}
\description{
Takes any number of 2D matrices, each with the same number of cols but with 
any set of named rows. Every col 1 of the matrices are combined into a new 
matrix, as are every col 2,3 etc.
}
\seealso{
merge_list_mx.by.rows
}

