% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{merge_list_mx.by.rows}
\alias{merge_list_mx.by.rows}
\title{Takes any number of 2D matrices, each with the same number of rows but with 
any set of named cols. Every row 1 of the matrices are combined into a new 
matrix, as are every row 2,3 etc.}
\usage{
merge_list_mx.by.rows(...)
}
\arguments{
\item{...}{matrices, or a list of matrices}
}
\value{
matrices, or a list of matrices after merged
}
\description{
Takes any number of 2D matrices, each with the same number of rows but with 
any set of named cols. Every row 1 of the matrices are combined into a new 
matrix, as are every row 2,3 etc.
}
