% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataframeUtil.r
\name{mergeAndRemoveKeyColumn}
\alias{mergeAndRemoveKeyColumn}
\title{Merges (ie: join) on a key column and returns merged rows, minus
the key column. Merging means selecting rows from key_column_name
that are in selected_keys, erroring if a row cannot be found.}
\usage{
mergeAndRemoveKeyColumn(df, key_column_name, selected_keys)
}
\arguments{
\item{df}{dataframe}

\item{key_column_name}{column to merge on}

\item{selected_keys}{values of keys in key_column_name to select}
}
\value{
a dataframe after merged.
}
\description{
Merges (ie: join) on a key column and returns merged rows, minus
the key column. Merging means selecting rows from key_column_name
that are in selected_keys, erroring if a row cannot be found.
}

