% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{meta.add.list.varname}
\alias{meta.add.list.varname}
\title{Appends a "varname" meta element to the meta of each element
in a list.}
\usage{
meta.add.list.varname(xlist, varnames = names(xlist))
}
\arguments{
\item{xlist}{list of objects}

\item{varnames}{parallel vector of varnames. Defaults to names(xlist).
If NULL, does nothing}
}
\value{
xlist with the corresponding varname append to the meta
 attribute of each xlist element.
}
\description{
Appends a "varname" meta element to the meta of each element
in a list.
}

