% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{modelVariableNames}
\alias{modelVariableNames}
\title{Return the variables used in the supplied terms.
Removes "LvlX" suffix, and surrounding "I( )"
as well as disaggregates squared variables}
\usage{
modelVariableNames(model, strip.Lvl = TRUE)
}
\arguments{
\item{model}{a glm model}

\item{strip.Lvl}{strip the LvlX from names. Defaults to TRUE.}
}
\value{
chr vector of variable names
}
\description{
Return the variables used in the supplied terms.
Removes "LvlX" suffix, and surrounding "I( )"
as well as disaggregates squared variables
}
