% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyProps.r
\name{modifyPropsAsBinLevels}
\alias{modifyPropsAsBinLevels}
\title{Takes a categorical var specified in separate binary level variables
and applies modifyProps and returns the result as a list of modified
binary level variables.}
\usage{
modifyPropsAsBinLevels(vecs.list, desiredProps, propens = NULL)
}
\arguments{
\item{vecs.list}{vectors to modify, eg: simframe[binLevelVarnames]}

\item{desiredProps}{desired proportions}

\item{propens}{propensities, if ANY}
}
\value{
same list of vectors, but with proportions modified
}
\description{
Takes a categorical var specified in separate binary level variables
and applies modifyProps and returns the result as a list of modified
binary level variables.
}

