% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{namedMatrix}
\alias{namedMatrix}
\title{Create a matrix of NA with specified col/row names/lengths.}
\usage{
namedMatrix(rows, cols)
}
\arguments{
\item{rows}{row names, or a numeric scalar for the number of rows}

\item{cols}{columns names, or a numeric scalar for the number of cols}
}
\value{
a named matrix of NA
}
\description{
Create a matrix of NA with specified col/row names/lengths.
}
