% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{normal.theory.CI.mean}
\alias{normal.theory.CI.mean}
\title{Called within normal.theory.CIs() to calculates a 95% confidence interval for proportions 
using assymptotic normal theory and replaces the first row of mx with it.  Called if 
statistic in normal.theory.CIs() is "mean".}
\usage{
normal.theory.CI.mean(dat, sg.var, mx)
}
\arguments{
\item{dat}{The data vector from which to calculate the confidence intervals}

\item{sg.var}{The variable defining the subgroup.  If there is no subgroup then this is a vector of 1s.}

\item{mx}{The matrix of collated results with confidence intervals.  The first row of this matrix will
be replaced.}
}
\value{
A matrix of the same dimension as mx with the first row replaced with assymnptotic normal 
theory 95% confidence intervals.
}
\description{
Called within normal.theory.CIs() to calculates a 95% confidence interval for proportions 
using assymptotic normal theory and replaces the first row of mx with it.  Called if 
statistic in normal.theory.CIs() is "mean".
}

