% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{normal.theory.CI.proportion}
\alias{normal.theory.CI.proportion}
\title{Called within normal.theory.CIs() to calculates a 95% confidence interval for proportions 
using assymptotic normal theory and replaces the first row of mx with it.  Called if 
statistic in normal.theory.CIs() is "proportion"}
\usage{
normal.theory.CI.proportion(dat, sg.var, mx, grpbyName, cat.adjustments,
  grping.logical)
}
\arguments{
\item{dat}{The data vector from which to calculate the confidence intervals}

\item{sg.var}{The variable defining the subgroup.  If there is no subgroup then this is a vector of 1s.}

\item{mx}{The matrix of collated results with confidence intervals.  The first row of this matrix will
be replaced.

#' @param grpbyName
variable name of a grouping variable

#' @param cat.adjustments
Used to determine whether "Not in subgroup"/"In subgroup" labels need to be used (used when
this function is being called in collate_all_run_results() and tables by subgroup are
being collated.

#' @param grping.logical
Takes the value TRUE or FALSE to indicate whether confidence intervals should be done 
separately for different subgroups.}
}
\value{
A matrix of the same dimension as mx with the first row replaced with assymnptotic normal 
theory 95% confidence intervals.
}
\description{
Called within normal.theory.CIs() to calculates a 95% confidence interval for proportions 
using assymptotic normal theory and replaces the first row of mx with it.  Called if 
statistic in normal.theory.CIs() is "proportion"
}

