% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{normal.theory.CIs}
\alias{normal.theory.CIs}
\title{Calculates a 95% confidence interval using assymptotic normal theory and replaces the first 
	#row of mx with it.}
\usage{
normal.theory.CIs(mx, statistic = "proportion", outcomes = NULL,
  simframe = NULL, cat.adjustments = NULL, grping.logical = TRUE,
  grpbyName = NULL, logiset = NULL, CI = TRUE)
}
\arguments{
\item{mx}{A matrix containing collated results with confidence intervals.}

\item{statistic}{if "proportion" (the default), it is assumed that confidence intervals for a proportion are 
desired otherwise, if "mean" it is assumed that confidence intervals for a mean are desired.}

\item{outcomes}{A list of matrices with each matrix containing the outcomes across all iterations (columns
correspond tpo iterations and rows to individuals).  Must be supplied.}

\item{simframe}{the simframe for the current environmenmt.  Used to get the data vector if the variable is 
not in outcomes.}

\item{cat.adjustments}{The cat.adjustments need to be passed to the function if a user specified subgroup has been
specified.  The logiset expression from the cat.adjustments will be used to define the 
subgroup variable and can also be used to get the binbreaks for continuous variables.}

\item{grping.logical}{Takes the value TRUE or FALSE to indicate whether confidence intervals should be done 
separately for different subgroups.}

\item{grpbyName}{variable name of a grouping variable}

\item{logiset}{a vector or matrix containing TRUEs and FALSEs. Confidence intervals 
will only be calculated for the TRUE subset of the data}
}
\value{
a matrix of the same dimension as mx with the first row replaced with assymnptotic normal 
theory 95% confidence intervals.
}
\description{
Calculates a 95% confidence interval using assymptotic normal theory and replaces the first 
	#row of mx with it.
}

