% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{nsort}
\alias{nsort}
\title{Sort numerically. 
If x can be represented numerically, then
sort numerically otherwise sort normally.}
\usage{
nsort(x, stripAlpha = TRUE, sortAlphaOnlySeparately = TRUE, ...)
}
\arguments{
\item{x}{object to sort}

\item{stripAlpha}{remove alpha characters before attempting sort, but retain them in the output}

\item{sortAlphaOnlySeparately}{if a value is purely alpha (ie: contains no numeric component) sort is separately
from the alpha-numeric values, and return it after any numerically sorted values}

\item{...}{additional arguments passed to sort}
}
\value{
x after sorted.
}
\description{
Sort numerically. 
If x can be represented numerically, then
sort numerically otherwise sort normally.
}
\seealso{
sort
}
