% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.r
\name{order_by_dlookup}
\alias{order_by_dlookup}
\title{Order the list by the results returned applying
dlookup to the list's elements}
\usage{
order_by_dlookup(dict, ...)
}
\arguments{
\item{dict}{a dictionary object}

\item{...}{objects to perform dlookup on}
}
\value{
... in order according to dlookup
}
\description{
Order the list by the results returned applying
dlookup to the list's elements
}
\seealso{
\code{\link{dlookup}}
}
