% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{predSimBinomsSelect_notChangeScores}
\alias{predSimBinomsSelect_notChangeScores}
\title{Predict and simulate binary value from 2 binomial models.}
\usage{
predSimBinomsSelect_notChangeScores(select, model0, model1,
  envir = parent.frame())
}
\arguments{
\item{select}{a logical vector, or vector of 0s and 1s, which determine
when to use model0 and when to use model1, i.e:
when select == 0, then result is predSimBinom using model0
when select == 1, then result is 1 - predSimBinom using model1}

\item{model0}{model0}

\item{model1}{model1}

\item{envir}{environment in which to evaluate model variables.}
}
\description{
Predict and simulate binary value from 2 binomial models.
}

