% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{predSimNormsSelect3Models}
\alias{predSimNormsSelect3Models}
\title{Predict and simulate value from 3 normal models.}
\usage{
predSimNormsSelect3Models(select, model1, model2, model3,
  envir = parent.frame())
}
\arguments{
\item{select}{a logical vector, or vector of 0,1,2,3which determine
when to use which model as follow,
when select == 0, then result is 0 
when select == 1, then result is predSimNorm using model1
when select == 2, then result is predSimNorm using model2
when select == 3, then result is predSimNorm using model3}

\item{model1}{model 1}

\item{model2}{model 2}

\item{model3}{model 3}

\item{envir}{environment in which to evaluate model variables.}
}
\description{
Predict and simulate value from 3 normal models.
}

