% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{remove.cols.named}
\alias{remove.cols.named}
\title{Remove cols by name.}
\usage{
remove.cols.named(x, cnames)
}
\arguments{
\item{x}{matrix or dataframe}

\item{cnames}{vector of colnames}
}
\value{
matrix after removed.
}
\description{
Remove cols by name.
}
