% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListUtil.r
\name{remove.elements}
\alias{remove.elements}
\title{Remove named elements from list}
\usage{
remove.elements(xlist, names_of_elements_to_remove)
}
\arguments{
\item{xlist}{list}

\item{names_of_elements_to_remove}{character vector}
}
\value{
the list after removed
}
\description{
Remove named elements from list
}
