% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r, R/Simmodule.r
\name{remove.NA.cols}
\alias{remove.NA.cols}
\title{Remove columns that contain all NAs.}
\usage{
remove.NA.cols(collated_results_freqs, num.runs)

remove.NA.cols(collated_results_freqs, num.runs)
}
\arguments{
\item{mx}{matrix}
}
\value{
matrix after removed.
}
\description{
Remove columns that contain all NAs.

remove.NA.cols
}
\examples{
mx <- matrix(c(1:4, NA, NA), nrow=2)
remove.NA.cols(mx)
}

