% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{remove.rows.named}
\alias{remove.rows.named}
\title{Remove rows by name.}
\usage{
remove.rows.named(mx, rnames)
}
\arguments{
\item{mx}{matrix or dataframe}

\item{rnames}{vector of rownames}
}
\value{
matrix after removed.
}
\description{
Remove rows by name.
}

