% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{remove.zero.rows}
\alias{remove.zero.rows}
\title{Remove rows that contain all zeros.}
\usage{
remove.zero.rows(mx)
}
\arguments{
\item{mx}{matrix}
}
\value{
matrix after removed.
}
\description{
Remove rows that contain all zeros.
}
