% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalAdjustment.r
\name{removeGlobalSubgroupFilterExpression}
\alias{removeGlobalSubgroupFilterExpression}
\title{Clear the subgroup expression for all cat.adjustments.}
\usage{
removeGlobalSubgroupFilterExpression(env.scenario)
}
\arguments{
\item{env.scenario}{Simenv object to be changed}
}
\value{
Simenv object
}
\description{
Clear the subgroup expression for all cat.adjustments.
}

