% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.r
\name{result.as.means.and.errs}
\alias{result.as.means.and.errs}
\title{Takes a result row and returns the means and error amounts as separate vectors in a matrix or list.}
\usage{
result.as.means.and.errs(result.row, simplify = T)
}
\arguments{
\item{result.row}{a result row, ie: a vector with values named Mean and Lower eg:

>  envs$`Scenario 1`$years1_5$run_results_collated$means$kids["Total",]
    Mean    Lower    Upper 
10.99488 10.62256 11.36721 

 if there are no values named Mean, then it will be assumed that all values
 are Means and that Lower is 0.}

\item{simplify}{if TRUE (default), returns a matrix instead of a list.}
}
\value{
a matrix/list of means and errs. The first row/means vector is the means from the result row, and the
 second row/errs vector is the difference between each mean and it's lower value.
}
\description{
Takes a result row and returns the means and error amounts as separate vectors in a matrix or list.
}
