% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalAdjustment.r
\name{setGlobalSubgroupFilterExpression}
\alias{setGlobalSubgroupFilterExpression}
\title{Set the subgroup expression to all cat.adjustments if the subgroup expression exists. 
Otherwise, remove the subgroup expression.}
\usage{
setGlobalSubgroupFilterExpression(env.scenario, subgroupExpression)
}
\arguments{
\item{env.scenario}{Simenv object to be changed}

\item{subgroupExpression}{the subgroup expression that is requested by the user. 
It specify the subgroup which is going to be adjusted.}
}
\value{
Simenv object
}
\description{
Set the subgroup expression to all cat.adjustments if the subgroup expression exists. 
Otherwise, remove the subgroup expression.
}
