% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simenv.r
\name{simulateP}
\alias{simulateP}
\title{Perform a simulation of X runs using parallel computing.}
\usage{
simulateP(Simenv, total_runs = 1)
}
\arguments{
\item{total_runs}{total number of runs to simulate}

\item{.}{Simenv receiving object}
}
\description{
NB: if it exists, uses propensities in global environment when doing adjustments for year 1
}
\examples{
\dontrun{
 . <- env.base
 env.base$simulate()
 . <- env.scenario
}
}

