% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{simulateRun}
\alias{simulateRun}
\title{Simulate outomes and store them in the outcomes list.}
\usage{
simulateRun(simenv, simulateFun)
}
\arguments{
\item{simenv}{simulation environment object}

\item{simulateFun}{simulation function input by the user}
}
\value{
a matrix contain the outcome
}
\description{
Sub-classes should extend this function.
}

