% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtil.r
\name{strip.alpha.mx}
\alias{strip.alpha.mx}
\title{Remove any alpha from row and col names, leaving only numbers.}
\usage{
strip.alpha.mx(mx)
}
\arguments{
\item{mx}{matrix}
}
\value{
matrix after modify the row and col names
}
\description{
Remove any alpha from row and col names, leaving only numbers.
}
