% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtil.r
\name{strip.alpha}
\alias{strip.alpha}
\title{Remove any alpha from character vector, leaving only numbers.}
\usage{
strip.alpha(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without alpha
}
\description{
Remove any alpha from character vector, leaving only numbers.
}

