% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtil.r
\name{strip.numeric}
\alias{strip.numeric}
\title{Remove any numbers from character vector, leaving only alpha.}
\usage{
strip.numeric(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character vector without numbers
}
\description{
Remove any numbers from character vector, leaving only alpha.
}
