% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{stripClass}
\alias{stripClass}
\title{Remove class attribute from an object}
\usage{
stripClass(x)
}
\arguments{
\item{x}{object}
}
\value{
x without "class" attribute
}
\description{
Remove class attribute from an object
}
