% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixUtil.r
\name{subsetFirstDimension}
\alias{subsetFirstDimension}
\title{Subset the first dimension, returning the same number of dimensions}
\usage{
subsetFirstDimension(x, logiset)
}
\arguments{
\item{x}{a vector, matrix or array}

\item{logiset}{vector to subset first dimension by}
}
\value{
subset of the vector, matrix or array
}
\description{
Subset the first dimension, returning the same number of dimensions
}
