% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtil.r
\name{trim}
\alias{trim}
\title{Remove leading and trailing spaces from a string}
\usage{
trim(string)
}
\arguments{
\item{string}{character vector}
}
\value{
character vector after remove spaces
}
\description{
Remove leading and trailing spaces from a string
}
