% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{tryerrorMsgs}
\alias{tryerrorMsgs}
\title{Return the messages of any element that is a try-error}
\usage{
tryerrorMsgs(xlist)
}
\arguments{
\item{xlist}{list}
}
\value{
vector that contains all components in the list with "try-error" class
}
\description{
Return the messages of any element that is a try-error
}
