% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simmodule.r
\name{adjustCatVar}
\alias{adjustCatVar}
\title{Adjust categorical values to desired proportions in cat.adjustments (if any).}
\usage{
adjustCatVar(
  x,
  varname,
  propens = NULL,
  desiredProps = NULL,
  simenv,
  iteration
)
}
\arguments{
\item{x}{categorical values to adjust}

\item{varname}{varname, used a lookup into cat.adjustments and propensities}
}
\description{
Allows subgroup adjustment if a subgroup expression attribute is attached to the cat.adjustments.
}
